[Rename] 컴포넌트 폴더 분리
- 컴포넌트를 level별로 분류하였습니다.
