+++ src/components/level-1/DiaryInput.tsx
